Add support for bookmark/link of numbers
Numbers may be bookmarked; as Word does not allow numbers to start a bookmark name, they are prefixed "no_" such as "no_1".
Links can be created to identical numbers. The same prefixing applies. When linking, the link point may also be prefixed
with the `^` character. This is useful for denoting text references. Any link point (not just numbers) may therefore be
prefixed with a `^`, and it will be ignored in matching to the correct bookmark as it is not a valid bookmark name character.